Split all cli commands up into separate source files
