+++ pkg/ecso/cli/environment_down.go
@@ -1,0 +1,38 @@
+package cli
+
+import (
+	"github.com/bernos/ecso/pkg/ecso"
+	"github.com/bernos/ecso/pkg/ecso/commands"
+	"github.com/bernos/ecso/pkg/ecso/config"
+	"github.com/bernos/ecso/pkg/ecso/dispatcher"
+	"gopkg.in/urfave/cli.v1"
+)
+
+func NewEnvironmentDownCliCommand(project *ecso.Project, dispatcher dispatcher.Dispatcher) cli.Command {
+	flags := struct {
+		Force cli.BoolFlag
+	}{
+		Force: cli.BoolFlag{
+			Name:  "force",
+			Usage: "Required. Confirms the environment will be terminated",
+		},
+	}
+
+	fn := func(ctx *cli.Context, cfg *config.Config) (ecso.Command, error) {
+		return makeEnvironmentCommand(ctx, project, func(env *ecso.Environment) ecso.Command {
+			return commands.NewEnvironmentDownCommand(env.Name, cfg.EnvironmentAPI(env.Region)).
+				WithForce(ctx.Bool(flags.Force.Name))
+		})
+	}
+
+	return cli.Command{
+		Name:        "down",
+		Usage:       "Terminates an ecso environment",
+		Description: "Any services running in the environment will be terminated first. See the description of 'ecso service down' for details. Once all running services have been terminated, the environment Cloud Formation stack will be deleted, and any DNS entries removed.",
+		ArgsUsage:   "ENVIRONMENT",
+		Action:      MakeAction(dispatcher, fn),
+		Flags: []cli.Flag{
+			flags.Force,
+		},
+	}
+}
