+++ pkg/ecso/cli/environment_rm.go
@@ -1,0 +1,38 @@
+package cli
+
+import (
+	"github.com/bernos/ecso/pkg/ecso"
+	"github.com/bernos/ecso/pkg/ecso/commands"
+	"github.com/bernos/ecso/pkg/ecso/config"
+	"github.com/bernos/ecso/pkg/ecso/dispatcher"
+	"gopkg.in/urfave/cli.v1"
+)
+
+func NewEnvironmentRmCliCommand(project *ecso.Project, dispatcher dispatcher.Dispatcher) cli.Command {
+	flags := struct {
+		Force cli.BoolFlag
+	}{
+		Force: cli.BoolFlag{
+			Name:  "force",
+			Usage: "Required. Confirms the environment will be removed",
+		},
+	}
+
+	fn := func(ctx *cli.Context, cfg *config.Config) (ecso.Command, error) {
+		return makeEnvironmentCommand(ctx, project, func(env *ecso.Environment) ecso.Command {
+			return commands.NewEnvironmentRmCommand(env.Name, cfg.EnvironmentAPI(env.Region)).
+				WithForce(ctx.Bool(flags.Force.Name))
+		})
+	}
+
+	return cli.Command{
+		Name:        "rm",
+		Usage:       "Removes an ecso environment",
+		Description: "Terminates an environment if it is running, and also deletes the environment configuration from the .ecso/project.json file",
+		ArgsUsage:   "ENVIRONMENT",
+		Action:      MakeAction(dispatcher, fn),
+		Flags: []cli.Flag{
+			flags.Force,
+		},
+	}
+}
