+++ pkg/ecso/cli/service_logs.go
@@ -1,0 +1,37 @@
+package cli
+
+import (
+	"github.com/bernos/ecso/pkg/ecso"
+	"github.com/bernos/ecso/pkg/ecso/commands"
+	"github.com/bernos/ecso/pkg/ecso/config"
+	"github.com/bernos/ecso/pkg/ecso/dispatcher"
+	"gopkg.in/urfave/cli.v1"
+)
+
+func NewServiceLogsCliCommand(project *ecso.Project, dispatcher dispatcher.Dispatcher) cli.Command {
+	flags := struct {
+		Environment cli.StringFlag
+	}{
+		Environment: cli.StringFlag{
+			Name:   "environment",
+			Usage:  "The environment to terminate the service from",
+			EnvVar: "ECSO_ENVIRONMENT",
+		},
+	}
+
+	fn := func(ctx *cli.Context, cfg *config.Config) (ecso.Command, error) {
+		return makeServiceCommand(ctx, project, func(service *ecso.Service, env *ecso.Environment) ecso.Command {
+			return commands.NewServiceLogsCommand(service.Name, env.Name, cfg.ServiceAPI(env.Region))
+		})
+	}
+
+	return cli.Command{
+		Name:      "logs",
+		Usage:     "output service logs",
+		ArgsUsage: "SERVICE",
+		Action:    MakeAction(dispatcher, fn),
+		Flags: []cli.Flag{
+			flags.Environment,
+		},
+	}
+}
